+++ utils/dsp.py
@@ -1,11 +1,13 @@
 import struct
 from pathlib import Path
-from typing import Dict, Any, Union
+from typing import Dict, Any, Union, List
 import numpy as np
 import librosa
+import torch
 import webrtcvad
-import soundfile as sf
 from scipy.ndimage import binary_dilation
+import torchaudio
+import torchaudio.transforms as transforms
 
 
 class DSP:
@@ -22,6 +24,7 @@ class DSP:
                  trim_start_end_silence: bool,
                  trim_silence_top_db:  int,
                  trim_long_silences: bool,
+                 target_dBFS: float,
                  vad_sample_rate: int,
                  vad_window_length: float,
                  vad_moving_average_width: float,
@@ -36,6 +39,7 @@ class DSP:
         self.n_fft = n_fft
         self.fmin = fmin
         self.fmax = fmax
+        self.target_dBFS = target_dBFS
 
         self.should_peak_norm = peak_norm
         self.should_trim_start_end_silence = trim_start_end_silence
@@ -47,64 +51,144 @@ class DSP:
         self.vad_moving_average_width = vad_moving_average_width
         self.vad_max_silence_length = vad_max_silence_length
 
+        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+
+        # init transformations
+        self.volume_transform = self.init_volume_transform()
+        self.mel_transform = self.init_mel_transform()
+
     @classmethod
-    def from_config(cls, config: Dict[str, Any]) -> 'DSP':
+    def from_config(cls, config: Dict[str, Any]) -> 'DSPTorchaudio':
+        """Initialize from configuration object"""
         return DSP(**config['dsp'])
 
-    def load_wav(self, path: Union[str, Path]) -> np.array:
-        wav, _ = librosa.load(str(path), sr=self.sample_rate)
-        return wav
+    def init_volume_transform(self):
+        """Initialize volume transformation"""
+        volume_transform = transforms.Vol(gain=self.target_dBFS, gain_type='db').to(self.device)
+        return volume_transform
 
-    def save_wav(self, wav: np.array, path: Union[str, Path]) -> None:
-        wav = wav.astype(np.float32)
-        sf.write(str(path), wav, samplerate=self.sample_rate)
-
-    def wav_to_mel(self, y: np.array, normalize=True) -> np.array:
-        spec = librosa.stft(
-            y=y,
+    def init_mel_transform(self):
+        """Initialize mel transformation"""
+        mel_transform = transforms.MelSpectrogram(
+            sample_rate=self.sample_rate,
             n_fft=self.n_fft,
+            win_length=self.win_length,
             hop_length=self.hop_length,
-            win_length=self.win_length)
-        spec = np.abs(spec)
-        mel = librosa.feature.melspectrogram(
-            S=spec,
-            sr=self.sample_rate,
-            n_fft=self.n_fft,
-            n_mels=self.n_mels,
-            fmin=self.fmin,
-            fmax=self.fmax)
-        if normalize:
-            mel = self.normalize(mel)
-        return mel
-
-    def griffinlim(self, mel: np.array, n_iter=32) -> np.array:
-        mel = self.denormalize(mel)
-        S = librosa.feature.inverse.mel_to_stft(
-            mel,
             power=1,
-            sr=self.sample_rate,
+            norm="slaney",
+            n_mels=self.n_mels,
+            mel_scale="slaney",
+            f_min=self.fmin,
+            f_max=self.fmax,
+        ).to(self.device)
+
+        return mel_transform
+
+    def load_wav(self, path: Union[str, Path], mono: bool = True) -> torch.Tensor:
+        """Load audio file into a tensor"""
+        effects = []
+        metadata = torchaudio.info(path)
+
+        # merge channels if source is multichannel
+        if mono and metadata.num_channels > 1:
+            effects.extend([
+                ["remix", "-"] # convert to mono
+            ])
+
+        # resample if source sample rate is different from desired sample rate
+        if metadata.sample_rate != self.sample_rate:
+            effects.extend([
+                ["rate", f'{self.sample_rate}'],
+            ])
+
+        waveform, _ = torchaudio.sox_effects.apply_effects_file(path, effects=effects)
+        return waveform
+
+    def save_wav(self, waveform: torch.Tensor, path: Union[str, Path]) -> None:
+        """Save waveform to file"""
+        torchaudio.save(filepath=path, src=waveform, sample_rate=self.sample_rate)
+
+    def adjust_volume(self, waveform: torch.Tensor) -> torch.Tensor:
+        """Adjust volume of the waveform"""
+        return self.volume_transform(waveform)
+
+    def adjust_volume_batched(self, data: List[torch.Tensor]) -> List[torch.Tensor]:
+        """Adjust volume of the waveforms in the batch"""
+        lengths = [tensor.size(1) for tensor in data]
+        padded_batch = [torch.nn.functional.pad(x, (0, max(lengths) - x.size(1))) for x in data]
+        stacked_tensor = torch.stack(padded_batch, dim=0)
+        processed_batch = self.adjust_volume(stacked_tensor)
+        result = [processed_waveform[:, :lengths[index]] for index, processed_waveform in enumerate(processed_batch)]
+        return result
+
+    def waveform_to_mel_batched(self, batch):
+        """Convert waveform to mel spectrogram for the batch of waveforms"""
+        lengths = [tensor.size(1) for tensor in batch]
+        expected_mel_lengths = [x // self.hop_length + 1 for x in lengths]
+        padded_batch = [torch.nn.functional.pad(x, (0, max(lengths) - x.size(1))) for x in batch]
+        batch_tensor = torch.stack(padded_batch, dim=0).to(self.device)
+        mels = self.waveform_to_mel(batch_tensor)
+        list_of_mels = [mel[:, :, :expected_mel_lengths[index]] for index, mel in enumerate(mels)]
+        return list_of_mels
+
+    def waveform_to_mel(self, waveform: torch.Tensor, normalized=True) -> torch.Tensor:
+        """Convert waveform to mel spectrogram"""
+        mel_spec = self.mel_transform(waveform)
+        if normalized:
+            mel_spec = self.normalize(mel_spec)
+        return mel_spec
+
+    def griffinlim(self, mel: Union[np.array, torch.Tensor], n_iter=32) -> np.array:
+        """Convert mel spectrogram to waveform using Griffin-Lim algorithm"""
+        if not torch.is_tensor(mel):
+            mel = torch.from_numpy(mel)
+            mel = torch.unsqueeze(mel, 0)
+            mel = mel.to(self.device)
+
+        mel = self.denormalize(mel)
+
+        inverse_melscale_transform = transforms.InverseMelScale(
+            n_stft=self.n_fft//2 + 1,
+            sample_rate=self.sample_rate,
+            f_min=self.fmin,
+            f_max=self.fmax,
+            mel_scale="slaney"
+        )
+
+        griffin_lim = transforms.GriffinLim(
             n_fft=self.n_fft,
-            fmin=self.fmin,
-            fmax=self.fmax)
-        wav = librosa.core.griffinlim(
-            S,
+            power=1,
             n_iter=n_iter,
-            hop_length=self.hop_length,
-            win_length=self.win_length)
-        return wav
+            win_length=self.win_length,
+            hop_length=self.hop_length
+        )
 
-    def normalize(self, mel: np.array) -> np.array:
-        mel = np.clip(mel, a_min=1.e-5, a_max=None)
-        return np.log(mel)
+        waveform = griffin_lim(inverse_melscale_transform(mel))
+        return waveform.numpy().squeeze(0)
 
-    def denormalize(self, mel: np.array) -> np.array:
-        return np.exp(mel)
+    def normalize(self, mel: torch.Tensor) -> torch.Tensor:
+        """Normalize mel spectrogram"""
+        mel = torch.clip(mel, min=1.e-5, max=None)
+        return torch.log(mel)
 
-    def trim_silence(self, wav: np.array) -> np.array:
-        return librosa.effects.trim(wav, top_db=self.trim_silence_top_db, frame_length=2048, hop_length=512)[0]
+    def denormalize(self, mel: torch.Tensor) -> torch.Tensor:
+        """Denormalize mel spectrogram"""
+        return torch.exp(mel)
+
+    def trim_silence(self, waveform: np.array) -> torch.Tensor:
+        """Trim silence from the waveform"""
+        trimmed_waveform = librosa.effects.trim(waveform,
+                                                top_db=self.trim_silence_top_db,
+                                                frame_length=self.win_length,
+                                                hop_length=self.hop_length)
+        trimmed_waveform = torch.from_numpy(trimmed_waveform[0])
+        trimmed_waveform = torch.unsqueeze(trimmed_waveform, 0)
+        return trimmed_waveform
 
     # borrowed from https://github.com/resemble-ai/Resemblyzer/blob/master/resemblyzer/audio.py
-    def trim_long_silences(self, wav: np.array) -> np.array:
+    def trim_long_silences(self, wav: Union[torch.Tensor, np.array]) -> np.array:
+        if torch.is_tensor(wav):
+            wav = wav.numpy().squeeze(0)
         int16_max = (2 ** 15) - 1
         samples_per_window = (self.vad_window_length * self.vad_sample_rate) // 1000
         wav = wav[:len(wav) - (len(wav) % samples_per_window)]
